Pull out the majority of functionality into a cog
