+++ py/matching.py
@@ -1,8 +1,10 @@
 """Utility functions for matchy"""
 import logging
+import discord
 from datetime import datetime, timedelta
 from typing import Protocol, runtime_checkable
-import state
+from state import State, save_to_file, ts_to_datetime
+import util
 import config
 
 
@@ -100,7 +102,7 @@ def get_member_group_eligibility_score(member: Member,
 
 
 def attempt_create_groups(matchees: list[Member],
-                          current_state: state.State,
+                          state: State,
                           oldest_relevant_ts: datetime,
                           per_group: int) -> tuple[bool, list[list[Member]]]:
     """History aware group matching"""
@@ -115,10 +117,10 @@ def attempt_create_groups(matchees: list[Member],
     while matchees_left:
         # Get the next matchee to place
         matchee = matchees_left.pop()
-        matchee_matches = current_state.get_user_matches(matchee.id)
+        matchee_matches = state.get_user_matches(matchee.id)
         relevant_matches = [int(id) for id, ts
                             in matchee_matches.items()
-                            if state.ts_to_datetime(ts) >= oldest_relevant_ts]
+                            if ts_to_datetime(ts) >= oldest_relevant_ts]
 
         # Try every single group from the current group onwards
         # Progressing through the groups like this ensures we slowly fill them up with compatible people
@@ -164,7 +166,7 @@ def iterate_all_shifts(list: list):
 
 
 def members_to_groups(matchees: list[Member],
-                      st: state.State = state.State(),
+                      state: State = State(),
                       per_group: int = 3,
                       allow_fallback: bool = False) -> list[list[Member]]:
     """Generate the groups from the set of matchees"""
@@ -176,14 +178,14 @@ def members_to_groups(matchees: list[Member],
         return []
 
     # Walk from the start of history until now trying to match up groups
-    for oldest_relevant_datetime in st.get_history_timestamps(matchees) + [datetime.now()]:
+    for oldest_relevant_datetime in state.get_history_timestamps(matchees) + [datetime.now()]:
 
         # Attempt with each starting matchee
         for shifted_matchees in iterate_all_shifts(matchees):
 
             attempts += 1
             groups = attempt_create_groups(
-                shifted_matchees, st, oldest_relevant_datetime, per_group)
+                shifted_matchees, state, oldest_relevant_datetime, per_group)
 
             # Fail the match if our groups aren't big enough
             if num_groups <= 1 or (groups and all(len(g) >= per_group for g in groups)):
@@ -198,3 +200,49 @@ def members_to_groups(matchees: list[Member],
     # Simply assert false, this should never happen
     # And should be caught by tests
     assert False
+
+
+async def match_groups_in_channel(state: State, channel: discord.channel, min: int):
+    """Match up the groups in a given channel"""
+    groups = active_members_to_groups(state, channel, min)
+
+    # Send the groups
+    for group in groups:
+
+        message = await channel.send(
+            f"Matched up {util.format_list([m.mention for m in group])}!")
+
+        # Set up a thread for this match if the bot has permissions to do so
+        if channel.permissions_for(channel.guild.me).create_public_threads:
+            await channel.create_thread(
+                name=util.format_list([m.display_name for m in group]),
+                message=message,
+                reason="Creating a matching thread")
+
+    # Close off with a message
+    await channel.send("That's all folks, happy matching and remember - DFTBA!")
+
+    # Save the groups to the history
+    state.log_groups(groups)
+    save_to_file(state)
+
+    logger.info("Done! Matched into %s groups.", len(groups))
+
+
+def get_matchees_in_channel(state: State, channel: discord.channel):
+    """Fetches the matchees in a channel"""
+    # Reactivate any unpaused users
+    state.reactivate_users(channel.id)
+
+    # Gather up the prospective matchees
+    return [m for m in channel.members if state.get_user_active_in_channel(m.id, channel.id)]
+
+
+def active_members_to_groups(state: State, channel: discord.channel, min_members: int):
+    """Helper to create groups from channel members"""
+
+    # Gather up the prospective matchees
+    matchees = get_matchees_in_channel(state, channel)
+
+    # Create our groups!
+    return members_to_groups(matchees, state, min_members, allow_fallback=True)
